Fix system-colour handling in rgb_from_ansi256
Six of the system colours returned by rgb_from_ansi256 were off by
a bit.  Only difference in colour 7 (light grey) was perceivable
and even then not by a much.  Also, the difference was in system
colours which aren’t standardised so the impact of this discrapency
is minimal if any.