Rewrite luminance-coefficients.py
Base RGB↔XYZ transformation matrices calculation on my blog post¹ which
makes the script natural companion to the entry.  In doing so, separate
implementation of standard matrix operations (namely multiplication and
inversion) and sRGB-specific constants to separate functions.  This way,
function implementing the calculation (calculate_rgb_matrix) becomes
short and easy to follow.

¹ https://mina86.com/2019/srgb-xyz-matrix/

While at it, make the luminance-coefficients.py compatible with
Python 2 (tested on Python 2.7) and 3 (tested Python 3.7) and rework
printing code so the output is much fancier and hopefully easier to
read and understand.

As a side effect white point calculation got fixed which improved
accuracy of the calculations.  In practice this doesn’t affect any
results but update lumimance benchmark with the new values nonetheless.