Fix Python 3.9 compatibility in luminance-coefficients.py
In effort to make the code compatible with Python 2.x and Python 3.x,
the script would use math.gcd if available and fractions.gcd
otherwise.  The way it was done, however, the code would always try to
access the latter even if the former is available.  In combination
with removal of fractions.gcd in Python 3.9, this caused the script to
break.

Fix by conditionally trying to access fractions.gcd if and only if
math.gcd is not present.

luminance-coefficients.py is an internal tool.  As such, this commit
does not affect the library code.