Port the algorithm to pure Rust
Having the code in Rust avoids the FFI boundary which makes various
optimisations impossible.  With the entire code written in Rust,
compiler can look into the algorithm and better optimise the code.

As a result, in benchmarks conversion from True Colour to ANSI index
sped up by 20% and conversion in the opposite direction by 65%.